requirements.txt
anyio==3.3.4
asgiref==3.4.1
beautifulsoup4==4.10.0
certifi==2021.10.8
charset-normalizer==2.0.7
click==8.0.3
fastapi==0.70.0
h11==0.12.0
httptools==0.2.0
idna==3.3
importlib-metadata==4.8.1
pydantic==1.8.2
pyrastreio==0.2.0
python-dotenv==0.19.1
PyYAML==6.0
Rastreio-Correios==0.1.4
requests==2.26.0
sniffio==1.2.0
soupsieve==2.3
starlette==0.16.0
tabulate==0.8.9
typing-extensions==3.10.0.2
urllib3==1.26.7
uvicorn==0.15.0
uvloop==0.16.0
watchgod==0.7
websockets==10.0
zipp==3.6.0
